+++ saved-sites/saved-sites-impl/src/main/res/drawable/ic_folder_24.xml
@@ -1,0 +1,26 @@
+<!--
+  ~ Copyright (c) 2022 DuckDuckGo
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M2,7c0,-2.21 1.79,-4 4,-4h4c1.657,0 3,1.343 3,3h5c2.21,0 4,1.79 4,4v7c0,2.21 -1.79,4 -4,4L6,21c-2.21,0 -4,-1.79 -4,-4L2,7zM6,5c-1.105,0 -2,0.895 -2,2v3h16c0,-1.105 -0.895,-2 -2,-2h-5c-1.105,0 -2,-0.895 -2,-2 0,-0.552 -0.448,-1 -1,-1L6,5zM20,12L4,12v5c0,1.105 0.895,2 2,2h12c1.105,0 2,-0.895 2,-2v-5z"
+      android:fillColor="?attr/daxColorPrimaryIcon"
+      android:fillType="evenOdd"/>
+</vector>
