+++ saved-sites/saved-sites-impl/src/main/res/drawable/ic_folder_add_24.xml
@@ -1,0 +1,29 @@
+<!--
+  ~ Copyright (c) 2022 DuckDuckGo
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M6,3C3.79,3 2,4.79 2,7v10c0,2.21 1.79,4 4,4h6c0.552,0 1,-0.448 1,-1s-0.448,-1 -1,-1L6,19c-1.105,0 -2,-0.895 -2,-2v-5h16v2c0,0.552 0.448,1 1,1s1,-0.448 1,-1v-4c0,-2.21 -1.79,-4 -4,-4h-5c0,-1.657 -1.343,-3 -3,-3L6,3zM20,10L4,10L4,7c0,-1.105 0.895,-2 2,-2h4c0.552,0 1,0.448 1,1 0,1.105 0.895,2 2,2h5c1.105,0 2,0.895 2,2z"
+      android:fillColor="?attr/daxColorPrimaryIcon"
+      android:fillType="evenOdd"/>
+  <path
+      android:pathData="M19,15c0,-0.552 -0.448,-1 -1,-1s-1,0.448 -1,1v2h-2c-0.552,0 -1,0.448 -1,1s0.448,1 1,1h2v2c0,0.552 0.448,1 1,1s1,-0.448 1,-1v-2h2c0.552,0 1,-0.448 1,-1s-0.448,-1 -1,-1h-2v-2z"
+      android:fillColor="?attr/daxColorPrimaryIcon"/>
+</vector>
