+++ saved-sites/saved-sites-impl/src/main/res/values/donottranslate.xml
@@ -37,10 +37,25 @@
     <string name="bookmarksSectionTitle">Bookmarks</string>
     <string name="addToFavorites">Add to Favorites</string>
     <string name="deleteBookmarkConfirmationDescription" instruction="Placeholder is the name of the bookmark">Are you sure you want to delete your bookmark for &lt;b>%1$s&lt;/b>?</string>
+    <string name="deleteFavoriteConfirmationDialogTitle" instruction="Dialog title for deleting a favorite site">Delete this favorite?</string>
+    <string name="deleteFavoriteConfirmationDialogDescription" instruction="Placeholder is the name of the favorite site">This will delete your favorite for &lt;b>%1$s&lt;/b>.</string>
     <string name="deleteBookmark" instruction="%1$s is the bookmark name">"Delete %1$s bookmark?"</string>
     <string name="deleteSavedSiteConfirmationDialogCancel">Cancel</string>
     <string name="deleteSavedSiteConfirmationDialogDelete">Delete</string>
     <string name="dialogConfirmTitle">Confirm</string>
     <string name="delete">Delete</string>
+    <string name="bookmarkDialogTitleEdit">Edit Bookmark</string>
+    <string name="favoriteDialogTitleEdit">Edit Favorite</string>
+
+    <!-- Edit Folders -->
+    <string name="folderLocationTitle">Select Location</string>
+    <string name="addFolder">Add Folder</string>
+    <string name="bookmarksActivityTitle">Bookmarks</string>
+    <string name="editFolder">Edit Folder</string>
+    <string name="deleteFolder" instruction="%1$s is the folder name">"Delete %1$s?"</string>
+    <plurals name="bookmarkFolderDeleteMessage">
+        <item quantity="one" instruction="%1$d is the number of items inside the folder">Are you sure you want to delete this folder and %1$d item?</item>
+        <item quantity="other" instruction="%1$d is the number of items inside the folder">Are you sure you want to delete this folder and %1$d items?</item>
+    </plurals>
 
 </resources>
